pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ogaden"]
include-package-data = true

[tool.setuptools.package-data]
ogaden = ["templates/*"]

[project]
name = "ogaden"
version = "0.1.0"
description = "Cryptocurrency trading bot for Binance spot markets"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "python-binance>=1.0.27",
    "pandas>=2.2.3",
    "pymemcache>=4.0.0",
    "ta>=0.11.0",
    "flask>=3.1.0",
    "flask-socketio>=5.5.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
]
server = [
    "gunicorn>=21.2.0",
    "gevent>=24.2.1",
    "gevent-websocket>=0.10.1",
]
all = ["ogaden[dev,server]"]

[project.scripts]
ogaden-engine    = "ogaden.engine:main"
ogaden-analysis  = "ogaden.analysis:main"
ogaden-dashboard = "ogaden.dashboard:main"


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.12"
strict = true
[[tool.mypy.overrides]]
module = "ogaden.indicators"
disable_error_code = ["attr-defined"]

[tool.ruff]
src = ["ogaden"]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
